+++ package/Editor/ContentPipeline.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections;
+using System.Collections.Generic;
 using System.IO;
 using UnityEditor.Build.Pipeline.Interfaces;
 using UnityEditor.Build.Pipeline.Utilities;
@@ -29,11 +31,33 @@ namespace UnityEditor.Build.Pipeline
         /// <returns>Return code with status information about success or failure causes.</returns>
         public static ReturnCode BuildAssetBundles(IBuildParameters buildParameters, IBundleBuildContent buildContent, out IBundleBuildResults result)
         {
+            var taskList = DefaultBuildTasks.Create(DefaultBuildTasks.Preset.AssetBundleCompatible, buildParameters.ScriptInfo == null);
+            return BuildAssetBundles(buildParameters, buildContent, out result, taskList);
+        }
+        
+        /// <summary>
+        /// Default implementation of generating Asset Bundles using the Scriptable Build Pipeline.
+        /// </summary>
+        /// <param name="buildParameters">Set of parameters used for building asset bundles.</param>
+        /// <param name="buildContent">Set of content and explicit asset bundle layout to build.</param>
+        /// <param name="result">Results from building the content and explicit asset bundle layout.</param>
+        /// <param name="taskList">Custom task list for building asset bundles.</param>
+        /// <returns>Return code with status information about success or failure causes.</returns>
+        public static ReturnCode BuildAssetBundles(IBuildParameters buildParameters, IBundleBuildContent buildContent, out IBundleBuildResults result, IList<IBuildTask> taskList)
+        {
             // Avoid throwing exceptions in here as we don't want them bubbling up to calling user code
             if (buildParameters == null)
             {
                 result = null;
                 BuildLogger.LogException(new ArgumentNullException("buildParameters"));
+                return ReturnCode.Exception;
+            }
+
+            // Avoid throwing exceptions in here as we don't want them bubbling up to calling user code
+            if (taskList.IsNullOrEmpty())
+            {
+                result = null;
+                BuildLogger.LogException(new ArgumentException("Argument cannot be null or empty.", "taskList"));
                 return ReturnCode.Exception;
             }
 
@@ -64,11 +88,10 @@ namespace UnityEditor.Build.Pipeline
                     BuildLogger.LogException(e);
                     return ReturnCode.Exception;
                 }
-
-                var pipeline = DefaultBuildTasks.Create(DefaultBuildTasks.Preset.AssetBundleCompatible, buildParameters.ScriptInfo == null);
-                exitCode = BuildTasksRunner.Validate(pipeline, buildContext);
+                
+                exitCode = BuildTasksRunner.Validate(taskList, buildContext);
                 if (exitCode >= ReturnCode.Success)
-                    exitCode = BuildTasksRunner.Run(pipeline, buildContext);
+                    exitCode = BuildTasksRunner.Run(taskList, buildContext);
 
                 if (Directory.Exists(buildParameters.TempOutputFolder))
                     Directory.Delete(buildParameters.TempOutputFolder, true);