+++ package/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.10-preview] - 2018-04-26
+- Added BuildAssetBundles API that takes a custom task list
+- Added null checks for BuildSettings.typeDB & AssetBundleBuild.addressableNames
+
 ## [0.0.9-preview] - 2018-04-04
 - Added documentation for IWriteOperation implementations
 - Added documentation for ReturnCodes, LegacyBuildPipeline, and ContentPipeline